Split debug system
